+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * chownloss operation:
   * dumped `chown` command since it's no longer needed; admins have to take care that sharing the files does not cause permission problems
   * removed -autoChown and -autoRechown options
+  * removed 'owner' status field
+  * stages are now startet as the 'created' user
+  * notify owner now sends emails to the creator
   
 * stool status: fixed exception is service wrapper has no child process